fix: Redis 자료형 에러 수정

+++ api-server/src/main/java/com/ticket/ticket/seat/SeatService.java
@@ -39,7 +39,7 @@ public class SeatService {
     }
 
     protected SeatPurchaseStatus buySeats(int seatNumber, String userId){
-        SeatPurchaseStatus status = redisRepository.buySeat(seatNumber);
+        SeatPurchaseStatus status = redisRepository.buySeat((long) seatNumber);
 
         if (status != SUCCESS){
             return status;
